+++ examples/test_zugbruecke.py
@@ -43,8 +43,7 @@ if any([platform.startswith(os_name) for os_name in ['linux', 'darwin', 'freebsd
 	f.write('{"log_level": 10}')
 	f.close()
 
-	import zugbruecke
-	ctypes = zugbruecke
+	import zugbruecke as ctypes
 
 elif platform.startswith('win'):